estel/requirements/dev-backend-requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-cors-headers==2.5.2
django-crontab==0.7.1
django-extensions==2.1.6
django-filter==2.2.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-geoposition-2==0.3.6
django-grappelli==2.13.1
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
googlemaps==3.0.2
icalendar==4.0.3
idna==2.8
ipdb==0.12.2
ipython==7.3.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
parso==0.3.4
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
Pygments==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
drf-writable-nested==0.5.1
django-ordered-model==3.3.0
